refactor: CouponManagement 로직 분리
- CouponManagement.tsx - 컨테이너 역할 (상태 관리, 탭/모달 제어)
- CouponTable.tsx - 쿠폰 목록 테이블
- CouponForm.tsx - 쿠폰 추가/수정 폼
