+++ src/basic/features/coupons/ui/index.ts
@@ -1,1 +1,3 @@
 export { CouponManagement } from './CouponManagement';
+export { default as CouponForm } from './CouponForm';
+export { default as CouponTable } from './CouponTable';
